Rename randomIntegerBetween -> internalRandomBetween in Scratch Block Library
internalRandomBetween should only be used within the Scratch Block Library
itself and NOT by client code. Instead, clients should call randomBetween,
which is an alias for internalRandomBetween.

The reason for this change is the following: when generating error witnesses
one has to create mocks for functions such as goToRandomPosition or
randomBetween. goToRandomPosition is implemented internally in the
Scratch Block Library by calling internalRandomBetween. Without the aliasing
it would not be possible to distinguish between calls to internalRandomBetween
that were made from the library or client code. The ability to do so is
important for creating mocks. These should only contain the relevant return
values and not some additional values that were used internally by the Scratch
Block Library.